feat(shared): scattered, more-vivid AI aurora on support chat (both apps)
Bring the Gemini-style gradient to the driver-app AI experience, which
lives in the shared SupportScreen chat tab (also used by rider-app):

- Add shared AiAuroraBackground: five blobs scattered across the screen
  (was three large washes), higher colour alpha and lower blur intensity
  so the red/orange/white reads more vivid / less transparent. Fades in
  on the idle chat and out once the rider/driver starts typing.
- Render it behind the SupportScreen chat tab, gated to the idle state
  (welcome message only, empty draft).
- Point rider-app's AiAuroraBackground at the shared component so the
  rider AI assistant inherits the same tuning from one source of truth.

Co-developed with Claude Code (claude-sonnet-4-6)
